+++ ui/carousel/src/components/Infinite.tsx
@@ -1,15 +1,14 @@
 import { CarouselContextProvider } from "../common/context"
 import type { CarouselProps } from "../common/types"
-import { swipePower } from "../common/utils"
+import { getSwipePower } from "../common/utils"
 import { arrayFromNumber } from "@social-hustle/utils-arrays"
 import cn from "@social-hustle/utils-classnames"
-import { useDimensions } from "@social-hustle/utils-hooks"
-import { getMin, getMax, clamp, wrap } from "@social-hustle/utils-numbers"
+import { useDimensions, useResizeObserver } from "@social-hustle/utils-hooks"
+import { clamp, wrap } from "@social-hustle/utils-numbers"
 import type { PanInfo } from "framer-motion"
 import { animate, motion, useMotionValue, useMotionValueEvent } from "framer-motion"
-import { Children, useEffect, useMemo, type CSSProperties, useReducer } from "react"
-import { useWindowSize, usePrevious } from "react-use"
-import { useDebouncedCallback } from "use-debounce"
+import { Children, useMemo, type CSSProperties, useReducer } from "react"
+import { useWindowSize } from "react-use"
 
 const getCurrentBreakpoint = (arr: number[], windowWidth: number) => {
   let breakpoint = 0
@@ -64,23 +63,13 @@ export interface InfiniteCarouselProps extends CarouselProps<RenderPropProps> {
   visibleItems?: VisibleItems
   /** Px value that needs to be exceeded to swipe.
    * TODO: Why is the default value 10000?
+   * @see https://codesandbox.io/p/sandbox/framer-motion-image-gallery-pqvx3?file=%2Fsrc%2FExample.tsx%3A28%2C1-28%2C4
    */
   swipePowerThreshold?: number
-  /** Integer between 0 and 1 denoting how much the drag offset has to satisfy the moveByPx value for a swipe to occur
-   * E.g. given a moveByPx of 500 and a swipeThreshold of 0.5, the drag offset has to be >= 250px to trigger a swipe
+  /** Integer between 0 and 1 denoting how much the drag offset has to satisfy the itemWidth value for a swipe to occur
+   * E.g. given a itemWidth of 500 and a swipeThreshold of 0.5, the drag offset has to be >= 250px to trigger a swipe
    */
   swipeThreshold?: number
-  /**
-   * Number of carousel items to translate on a swipe
-   * @default 1
-   */
-  moveBy?: number
-  /** Value to debounce the drag transition by.
-   * E.g. if debounceBy is 200, the carousel will debounce the drag transition by 200ms.
-   * NOTE: this is only used when dragging the carousel, not when controlling the carousel via the setPage function.
-   * @default 200
-   */
-  debounceBy?: number
   /** The starting page/index. Defaults to 0 */
   startAt?: number
 }
@@ -107,12 +96,10 @@ export const InfiniteCarousel = ({
   itemClassName = "",
   visibleItems = 2,
   swipePowerThreshold = 10000,
-  swipeThreshold = 0.9,
-  moveBy = 1,
+  swipeThreshold = 0.2,
   renderAfter,
   renderBefore,
   dragProps,
-  debounceBy = 200,
   draggable = true,
   startAt = 0,
   centered = false,
@@ -120,16 +107,20 @@ export const InfiniteCarousel = ({
   const childrenArray = Children.toArray(children)
   const childCount = Children.count(children)
 
-  const [ref, { width }] = useDimensions<HTMLDivElement>()
+  const [ref, { width }, node] = useDimensions<HTMLDivElement>()
 
   const visibleItemsNumber = useVisibleItems(visibleItems)
   const itemWidth = width / visibleItemsNumber
 
+  /**
+   * If `centered: true`, items should be offset so the carousel is centered
+   */
   const offsetFactor = centered ? Math.floor(visibleItemsNumber / 2) : 0
-  const offset = offsetFactor * itemWidth
 
-  /** Pixel value to translate the carousel */
-  const moveByPx = itemWidth * moveBy
+  /**
+   * Calculate `offsetFactor` by `itemWidth` to get the value in px
+   */
+  const offset = offsetFactor * itemWidth
 
   const [{ page, dragging }, dispatch] = useReducer(
     (state: State, action: Action): State => {
@@ -143,7 +134,7 @@ export const InfiniteCarousel = ({
           return {
             ...state,
             page: action.page,
-            //dragging: false,
+            dragging: false,
           }
         default:
           return state
@@ -155,9 +146,9 @@ export const InfiniteCarousel = ({
     }
   )
 
-  const prepend = arrayFromNumber(visibleItemsNumber).map((i) => 0 - i - 1)
+  const prepend = arrayFromNumber(childCount).map((i) => 0 - i - 1)
 
-  const append = arrayFromNumber(visibleItemsNumber).map((i) => childCount + i)
+  const append = arrayFromNumber(childCount).map((i) => childCount + i)
 
   const loopedChildren = useMemo(() => {
     const childMap = new Map<number, React.ReactNode>()
@@ -183,66 +174,64 @@ export const InfiniteCarousel = ({
 
   const setDragging = (dragging: boolean) => dispatch({ type: "SET_DRAGGING", dragging })
 
-  const min = -visibleItemsNumber + offsetFactor
-  const max = childCount - offsetFactor
-
   const setPage = (page: number) => {
     dispatch({ type: "SET_PAGE", page })
-    const animateTo = -page * moveByPx
-    void animate(x, animateTo).then(() => {
-      if (page === max) {
-        console.log("max")
-        x.set(0)
-        dispatch({ type: "SET_PAGE", page: 0 })
-      }
-      if (page === min) {
-        console.log("min")
-        const resetToPage = childCount - visibleItemsNumber
-        x.set(-resetToPage * moveByPx)
-        dispatch({ type: "SET_PAGE", page: resetToPage })
+  }
+
+  const setAnimation = async (value: number) => {
+    const animateTo = -value * itemWidth
+    return await animate(x, animateTo)
+  }
+
+  const translateCarousel = (newPageNumber: number) => {
+    setPage(newPageNumber)
+    const newWrappedPageNumber = wrap(newPageNumber, [0, childCount])
+    void setAnimation(newPageNumber).then(() => {
+      if (newPageNumber !== newWrappedPageNumber) {
+        x.set(-newWrappedPageNumber * itemWidth)
+        setPage(newWrappedPageNumber)
       }
     })
   }
 
-  const calculateDragConstraints = () => {
-    const left = (page + moveBy) * -moveByPx
-    const right = (page - moveBy) * -moveByPx
-    return { left, right }
+  const dragConstraints = {
+    left: -itemWidth * append.length - page * itemWidth + offset,
+    right: itemWidth * prepend.length - page * itemWidth - offset,
   }
 
   /**
    * The threshold in pixels that needs to be exceeded to trigger a `swipe`, which will move the carousel to the next page.
-   * This is calculated by multiplying the `moveByPx` value by the `swipeThreshold` prop.
-   * E.g. if `moveByPx` is 500 and `swipeThreshold` is 0.5, the `swipePxThreshold` will be 250.
+   * This is calculated by multiplying the `itemWidth` value by the `swipeThreshold` prop.
+   * E.g. if `itemWidth` is 500 and `swipeThreshold` is 0.5, the `swipePxThreshold` will be 250.
    * This means that the user has to drag the carousel by >= 250px to trigger a swipe.
    */
-  const swipePxThreshold = moveByPx * clamp(swipeThreshold, [0, 1])
+  const swipePxThreshold = itemWidth * clamp(swipeThreshold, [0, 1])
 
-  const handleEndDrag = useDebouncedCallback(
-    (e: Event, dragProps: PanInfo) => {
-      const { offset, velocity } = dragProps
-      const swipe = swipePower(offset.x, velocity.x)
+  const calculatePage = (offsetX: number) => {
+    const newPage =
+      offsetX > 0 ? Math.floor(offsetX / -itemWidth) + page : Math.ceil(offsetX / -itemWidth) + page
 
+    return newPage
+  }
+
+  const handleEndDrag = (e: Event, dragProps: PanInfo) => {
+    const { offset, velocity } = dragProps
+    const swipePower = getSwipePower(offset.x, velocity.x)
+
+    const translateTo = calculatePage(offset.x)
+
+    const shouldSwipeBasedOnDistance = Math.abs(offset.x) > swipePxThreshold
+    const shouldSwipeBasedOnPower = Math.abs(swipePower) > swipePowerThreshold
+    const shouldSwipe = shouldSwipeBasedOnDistance || shouldSwipeBasedOnPower
+
+    if (shouldSwipe) {
+      translateCarousel(translateTo)
+    } else {
       setDragging(false)
-
-      // If dragging RTL
-      if (offset.x < -swipePxThreshold || swipe < -swipePowerThreshold) {
-        setPage(page + moveBy)
-
-        // If dragging LTR
-      } else if (offset.x > swipePxThreshold || swipe > swipePowerThreshold) {
-        setPage(page - moveBy)
-      } else {
-        // If the user didn't drag far enough to trigger a swipe, animate the carousel back to original position
-        const animateTo = -page * moveByPx
-        void animate(x, animateTo)
-      }
-    },
-    debounceBy,
-    {
-      leading: true,
+      // If the user didn't drag far enough to trigger a swipe, animate the carousel back to original position
+      void setAnimation(page)
     }
-  )
+  }
 
   const handleDragStart = () => {
     setDragging(true)
@@ -253,16 +242,19 @@ export const InfiniteCarousel = ({
   const isDisabled = childCount < visibleItemsNumber || !draggable
 
   const renderProps: RenderPropProps = {
-    setPage,
+    setPage: translateCarousel,
     page,
   }
 
-  /**
-   * If width changes, reset the page to the startAt value.
-   */
-  useEffect(() => {
-    setPage(startAt)
-  }, [startAt, width])
+  useResizeObserver(node, () => {
+    translateCarousel(page)
+  })
+
+  /**useMotionValueEvent(x, "change", (latest) => {
+    const pageNumber = Math.round(latest / -itemWidth)
+    const wrappedPage = wrap(pageNumber, [0, childCount])
+    console.log({ wrappedPage })
+  })**/
 
   return (
     <CarouselContextProvider
@@ -298,10 +290,11 @@ export const InfiniteCarousel = ({
           dragElastic={0}
           onDragEnd={handleEndDrag}
           onDragStart={handleDragStart}
+          //onDrag={(event, info) => console.log({ event, info })}
           style={{
             x,
           }}
-          dragConstraints={calculateDragConstraints()}
+          dragConstraints={dragConstraints}
         >
           {loopedChildren.map(([i, child], index) => {
             return (