+++ ui/carousel/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @social-hustle/ui-carousel
 
+## 6.3.0
+
+### Minor Changes
+
+- fix animation bugs, optimize animations & renders
+
+## 6.2.0
+
+### Minor Changes
+
+- fix infinite carousel looping
+
 ## 6.1.0
 
 ### Minor Changes